remove old config storybook
