--- packages/web/.storybook/config.js
@@ -1,6 +1,0 @@
-import { addDecorator } from '@storybook/react';
-import themeDecorator from './theme-decorator';
-import routerDecorator from './router-decorator';
-
-addDecorator(themeDecorator);
-addDecorator(routerDecorator);
